Major Change: Change UI and Plugin Architecture
Previously the UIs and Plugins did not work well when
there was more than one videoplayer on a page.

This should now be fixed, due to the fact that the videoplayer itself
now calls `new Plugin()` on `initPlugin()` and
`new UI()` on `initUI()` for each video player instance.

But that means, that all the Plugins and UIs need to be updated as well,
to work with the current implementation.

The currently compatible plugins and UIs are:

 - IVAD Plugin 1.0.0
 - IVAD UI 1.0.0
 - Endcard Plugin 2.0.0
 - Livestream UI 1.0.0