Update hls.js + Always load and use hls.js
Updated hls.js to latest version.
Also always use hls.js if supported in the browser.

Previously we checked if the browser supports playing hls
natively, then we proceeded to play it natively and only used
hls.js as a kind of fallback.

Now, we check if the browser supports hls.js and if so,
we use it right away.

We use native functions of the browser only as fallback,
if hls.js is not supported in the browser.

This should also fix
https://github.com/stroeer/video/issues/24

Probably fixed `bufferStalledError` too.