🐛 Load `<source>` via `data-src` attribute
Previously we had a html markup like this:

```
<video>
  <source src="https://example.com/manifest.m3u8">
</video>
```

This caused native players to immediately start fetching the manifest
file. Which resulted in a race-condition on Google Chrome on Android on
t-online.de.

When Google Chrome on Android on t-online fetched the manifest before
our js (hls.js) kicked in, everything broke down, because the request
was cached and the cached request (made via the native browser
functionality) had no cors header present.

Now, we load an empty `<video>` tag with a `data-src` attached to it,
like so:

```
<video data-src="https://example.com/manifest.m3u8"></video>
```

This prevents the native browsers to start preloading the manifest file.

This is pretty much 100% the approach that is shown in the example of
hls.js:

https://hls-js.netlify.app/demo/basic-usage

```

<html>
  <head>
    <title>Hls.js demo - basic usage</title>
  </head>

  <body>
    <script src="../dist/hls.js"></script>

    <center>
      <h1>Hls.js demo - basic usage</h1>
      <video height="600" id="video" controls></video>
    </center>

    <script>
      var video = document.getElementById('video');
      if (Hls.isSupported()) {
        var hls = new Hls({
          debug: true,
        });
        hls.loadSource('https://test-streams.mux.dev/x36xhzz/x36xhzz.m3u8');
        hls.attachMedia(video);
        hls.on(Hls.Events.MEDIA_ATTACHED, function () {
          video.muted = true;
          video.play();
        });
      }
      // hls.js is not supported on platforms that do not have Media Source Extensions (MSE) enabled.
      // When the browser has built-in HLS support (check using `canPlayType`), we can provide an HLS manifest (i.e. .m3u8 URL) directly to the video element through the `src` property.
      // This is using the built-in support of the plain video element, without using hls.js.
      else if (video.canPlayType('application/vnd.apple.mpegurl')) {
        video.src = 'https://test-streams.mux.dev/x36xhzz/x36xhzz.m3u8';
        video.addEventListener('canplay', function () {
          video.play();
        });
      }
    </script>
  </body>
</html>
```